+++ Curso Udemy/Atividade/Composição/src/entities/Product.java
@@ -1,14 +1,16 @@
 package entities;
 
+import java.math.BigDecimal;
+
 public class Product {
 
     private String name;
-    private Double price;
+    private BigDecimal price;
 
     public Product() {
     }
 
-    public Product(String name, Double price) {
+    public Product(String name, BigDecimal price) {
         this.name = name;
         this.price = price;
     }
@@ -21,11 +23,11 @@ public class Product {
         this.name = name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
